CFY-6380 Use `C:\Program Files\Cloudify Agents` as base directory
The agent is still installed in a subdirectory using the machine name.
However, in the base directory the installer and the custom environment
file are left, so it's cleaner to have a `Cloudify Agents` directory.

For the future, the installer should be downloaded to a temporary file
and the environment file should probably be moved to the work
subdirectory in the agent environment.