CFY-7124 Fix custom env not being sourced by the daemon (#285)
* CFY-7124 Fix custom env not being sourced by the daemon

* Remove the shebang from custom env files

The env files are sourced using `source file` or `. file`, so the
shebang (#!/bin/bash) isn't needed - they are not executed.

If they were executed, the env woudln't be persisted in the
outside shell, anyway, so they would be useless then.

The shebang being there does nothing, only suggests incorrect usage -
that the files can be executed, but they would do nothing then.