Get remote temporary directory
Before this change, since `.mktemp` creates a file with the `.tmp`
extension instead of the desired `.ps1`, a suffix was added to the
created temporary name. However, the result of this is that two files
are created, that is, the call to `.delete` leaves one file in the
remote machine.

After this change, the temporary directory is used to get the path to
the remote file based on the original script filename. This way, only
one file is created remotely and only one call to `.delete` is needed.