CFY-7142. Fix command line too long (#289)
* Update indentation to improve readability

* Replace chained calls with a for loop

This way there's no need to write `.replace` multiple times

* Split content into chunks to avoid command line too long error

Maximum allowed commmand line length should be 2047 in old windows
instances:
https://support.microsoft.com/en-us/help/830473/command-prompt-cmd--exe-command-line-string-limitation

* Update docstring

* Add missing commas to list as separator

* Use single quotes for chunks

This reduces the amount of escaping needed

* Split script by lines

The `Add-Content` cmdlet adds a new line on its own when executed, so
the easiest way to split the contents is to do it by line.

For the future, multiple lines can be aggregated to reduce the amount of
calls to `Add-Content`.

* Escape single quotes properly

This is needed because in the `Add-Content` call the script is now
quoted with single quotes.

* Add quotes around package URL

This is not needed right now, but it is good to have it at least for
consistency.

* Preserve .ps1 extension so that script can be executed

* Run script through powershell

* Get remote temporary directory

Before this change, since `.mktemp` creates a file with the `.tmp`
extension instead of the desired `.ps1`, a suffix was added to the
created temporary name. However, the result of this is that two files
are created, that is, the call to `.delete` leaves one file in the
remote machine.

After this change, the temporary directory is used to get the path to
the remote file based on the original script filename. This way, only
one file is created remotely and only one call to `.delete` is needed.

* Split contents into chunks made of lines

Chunks must be made of lines because `Add-Content` adds a new line on
its own for each call.

* Remove duplicated comment

* Add split into chunks test cases

Also, fix implementation for edge cases

* Update to be python 2.6 compatible