CFY-7105. Refactor agent installation script templates (port to 4.2) (#299)
* CFY-7105 Update linux agent installation script

Separate install, configure and start parts. This way when there's no
need to have all of them it can be removed when the template is
rendered.

* CFY-7105 Update windows agent installation script

Separate install, configure and start parts. This way when there's no
need to have all of them it can be removed when the template is
rendered.

* CFY-7105 Include install, configure and start parts by default

The logic to decide when any of those is false will be added later.

* CFY-7105 Fix if condition

* CFY-7105 Remove whitespace when rendering the install script

This will result in a more readable script which makes easier to
troubleshoot problems.

* CFY-7105 Use here document to render custom environment script

No need to append to the file one by one, when the whole content of the
script can be rendered to a here document that is written in one step.

* CFY-7105 Append all commands to to the script at once

No need to write each command one by one

* CFY-7105 Add main fuction to linux/windows templates

This is to make easier to executed test cases given that the tests
assume that only one function call will happen just at the end of the
script.

* CFY-7105 Remove duplication function

This should have been removed during the rebase process

* CFY-7105 Export configure_agent function