On an error is raised while creating a runner, retry and don't continue (#311)
Before the refactor, we used to run .retry() and not continue trying
to use the runner. Now we need to make sure this behaviour is kept:
we set .retry(), and then we need to make sure no further attempts
to use the runner are done. So, in that case, we simply `return` from
the caller function.