CFY-7345 Make sure a `pip install` error is rethrown
We cannot allow the CommandExecutionError to bubble up, because then
the operation will be (hopelessly) retried. Instead, we should re-throw
it as an PluginInstallationError so that the framework above us changes
it into an NonRecoverableError and stops trying.

We also log the whole output, because for the common failures, the
reason can be seen there easily (eg. not having gcc when trying to
install the openstack plugin from source).