Run tox under tini
We create subprocesses in the tests (eg. detach agent) that need to
be reaped by an init system, or otherwise they become zombies and
fail the tests because the agent status commands thinks the process
is still running.

We can't run an init system directly on circleci (apparently, using
an image with an ENTRYPOINT set to tini doesn't work), so we can
only run tini as a subreaper and run tox under it.