Precreate the service queue before the agent starts
That way, when the agent does start, it already sees the ping message.

Otherwise, the first ping message is a race between sending the
message, and the agent starting and creating the queue - if the agent
loses that race, the message will be routed to nothing and lost, and
the agent will have to wait for the next ping retry. (in practice,
the agent always loses that race)

This means all agent starts will be faster by about 3 seconds