Agent install script: patch up cluster_config_path
For silly reasons (that have been fixed in 4.5.5 as CY-776) agents
are installed with an empty variable for the cluster config path.

This changes the install script, to patch up the config file, to
contain the correct value for that path. This, surprisingly, doesn't
come at a loss of generality, because the same paths that we
half-hardcode here, are also half-hardcoded in the python side.

This doesn't affect non-cluster agents, because the cluster module
will also check if the file exists, before trying to use it