create-agent-package script: specialcase the python
This is a hack :(
Depending on the version, we tell the agent-packager to use that
version as well.
I'm not overly worried about this, because the scenario isn't super
realistic, with creating the agent package inside a script-runner task.

Hopefully we'll be able to get rid of this, if we are able to
fix the agent-packager.