rewrite test_install_from_wagon_concurrent
it had to be rewritten because now the logger mock is different,
because we use ctx.logger

it's also a bit stricter now, because now it requires thread-safety
(by using threads), which is a stronger requirement than just
multiprocess-safety. Also now it can be run on windows.

..but i mostly changed this to threads for the logging check
to be less insane