CY-2658 Make a virtualenv when installing plugins (#617)
* get_*_path utils: parametrize virtualenv

those are useful but would be more useful if they were able to use
virtualenvs

* plugin installer: create a full virtualenv

Instead of making a sub-virtualenv, "just library dirs", etc - make
a proper virtualenv and install stuff there. Simple as that.

- the "constraints" idea is gone because it's an anti-feature - it
  used to be so that libs that are in the agent wouldn't be installed,
  but the point is that yes, they should be
- the tmpdir idea is gone as well because a virtualenv must be made
  wherever it is actually going to be used (relocatable is not
  supported enough)
- uninstall does just `rm -rf` anyway so no changes needed there

* PluginInstaller: drop the class

no reason for this to be a class, it's just an ad hoc collection
of methods

* use ctx.logger for logging

using a global logger doesn't actually pass the logs through (to
the database etc)

also replace str.format with %s and logging args

that first log there i'm just removing because it's a debug, so
nobody needed it, or saw it, in years; and i'm not going to do
the platform stuff just for a log which isn't even shown

* update plugin installer usage in unittests

there's no PluginInstaller class, so we don't need to instantiate it,
obviously. Just call the functions.

* tests: mock CloudifyContext

that is the right context; the other one is just a dict. This is
the actual context mock.

* link the newly-created venv

make the new venv also import from the agent venv (pth files are
appended to the path, after venv-local packages)

also, copy agent's .pth files content as well - currently, the
only way the agent would _have_ a pth file, is when it installs
something with --editable, which only ever happens when done
manually (by me, while debugging). But might as well handle that.

this is so that plugins who do not depend on cloudify-common can
still work: they will just use common from the agent venv

* drop useless test

'versions conflict' doesn't happen anymore because well, separate venv

* rewrite test_install_from_wagon_concurrent

it had to be rewritten because now the logger mock is different,
because we use ctx.logger

it's also a bit stricter now, because now it requires thread-safety
(by using threads), which is a stronger requirement than just
multiprocess-safety. Also now it can be run on windows.

..but i mostly changed this to threads for the logging check
to be less insane

* remove unused functions

* docstrings

* Move get_*_path to cloudify-common

There's a similar function in common, so let's use that

* Make the virtualenv without pip, wheel, setuptools

This saves on the order of 15MB per plugin