Remove the crontab entry in before_self_stop
We've had this `before_self_stop` hook for years and it's never
really been used for much; it seems it's a good place to put
the cron stopping stuff.

Directly, this is because there's two ways of stopping the agent:
the stop method (called from the cfy-agent cli), and the stop_amqp
operation, which just raises StopAgent.
Both of those approaches need to do the same cleanup (ie. crontab),
so it's nicer to have the cron stuff in a separate method.