RD-2618 Move creating operation subprocesses to worker.py (#800)
* Move subprocess-related things from dispatch to here

* process-registry: use execution-id as the key

it's already the key, and it's used like that directly in .cancel(),
so let's just use it directly in register/unregister as well.

The make_key was back when we didnt know what we wanted to actually
be the key. Now we know that it's the execution id. And the handler
no longer has a .ctx attribute.

* fixup print_task

now that it takes the actual ctx as a parameter, it must handle the
various types of ctx

* make registry be optional after all

used to be before

* move handling subprocess output to a separate method

* add a few tests

* use os.path.join instead of pathlib

old python compat

* fix a typo in this hint

* test_installer: rewrite _assert_task_runnable

instead of the full dispatch thing, just have a mock small
implementation with a single subprocess call.
This is because `dispatch.dispatch()` no longer calls a subprocess
if task_target is set (which was the case _only_ in these tests;
the only other time it's called, is from cfy local, where that
is not set)

* dev-reqs

* loosen the assert here

unfortunate. Comment explains it.

* set -ex for the tests

can't set -u because without things like PS1 etc it breaks the
activate/deactivate

and not having activate breaks the tests

pffff

* Revert "dev-reqs"

This reverts commit de60f169be75616fb3e7776d42b213b4347397c5.