RD-6658 Use `cfy-agent setup` for creating an agent (#867)
* Base Daemon: take all params as arguments

Instead of having a dict, just take the things as arguments. Those
arguments can also be mandatory, and so, a bunch of "validate" methods
is just unnecessary.

* Use daemon.as_dict instead of the internal method

The .as_dict was added in the previous commit, and to remind you, is:
```python
def as_dict(self):
    return {k: v for k, v in vars(self).items() if not k.startswith('_')}
```

Now look at the old implementation. What the hell.

* New linux.sh.template

This script will now download & install the agent.

* New windows.ps1.template

Similar to linux.

* Remove the -download.template scripts

They're unnecessary anymore.

* Implement `cfy-agent setup`

The scripts call this CLI, so let's actually implement it!
It pulls the node-instance settings, and creates an agent.

* installer: stop using the long list of envvars

now, in the scripts, all that is needed is just the arguments
passed explicitly to `cfy-agent setup`. That whole envvars stuff
is no longer required.

* create the agent record when preparing agent config

instead of JUST updating runtime-props, also create the agent
record right here.
This means non-remote agents will now have their record created up
front as well, and so `cfy-agent setup` can actually query
this info.

Before, we updated runtime-props first, then installed the
agent, and only then created the record. That was stupid.

* update process management config templates

Now, extra_env is rendered right there, instead of in a separate file.
Thanks to this, .extra_env_path is just not required anymore

* daemons delete: just stop the agent

Look, I've really had it with this force etc. bullshit. Delete just
stops the agent, that's it. If it can't, tough luck.

* remove useless agent_config stuff

these things are now just unused. Most of them we set in the script.

* update script template renderer

No need for separate init_script/install_script now, but let's leave
both functions in (they do the same thing) to keep back-compat.

* update daemon args

The point is, we don't store rest_host anymore. That is always passed
in with an operation message.

* remove disable-requiretty.sh

it is now inlined into the main linux.sh script

* update daemons CLI

- create now, well, creates
- configure no longer exists

* A LOT of test updates

Look, this is pretty boring. I drop several tests, too. Some of them
are just too tightly tied to the old implementation, and now are
useless.

At the end of the day, this is only really tested in inte-tests and
sys-tests...

* unpin unnecessary pins

* remove pointless -e in shebang

Anyway, the shebang is not all that useful. This script is USUALLY
going to be called with `bash` explicitly. But it doesn't have to be.

* no need for g+w, actually!

The agent never writes to that directory now.