Agent setup: pass node instance id separately
Agent name and node instance id are not ALWAYS the same:
in an upgrade, the agent name has an uuid appended to it,
but node instance id stays the same.

Pass it separately so that upgraded agents can still find their
instance.

Note: it is already set in conf
(in `CloudifyAgentConfig.set_default_values`)