RD-7005 Agent upgrade directory fixups (#883) (#886)
* Agent setup: pass node instance id separately

Agent name and node instance id are not ALWAYS the same:
in an upgrade, the agent name has an uuid appended to it,
but node instance id stays the same.

Pass it separately so that upgraded agents can still find their
instance.

Note: it is already set in conf
(in `CloudifyAgentConfig.set_default_values`)

* RD-7005 agent upgrade: keep old basedir

If the user set a basedir before, we should keep it after upgrade
as well. Can't just revert to the default, that'll break all the
tests that use detach at least.

* Install the agent into install_dir

Now, separate .basedir from .install_dir - the agent is
going to be installed into the provided basedir, and if it's
not provided, only then default the install_dir to a versioned
directory...

This way, in agent upgrades, basedir can be just copied from the
old agent

* drop unused create_custom_env_file

custom env is already put in the regular env file, not in a separate
custom env file. So this is unnecessary and unused. Which, funnily
makes the whole RemoteInstallerMixin unused as well

* update test_create_agent_dict

now we do expect those to be the same, because now we consider
the basedir to be provided by the user in this test

* test_configuration: handle the new attrs