RND-262 Upgrade operation: put rest_port in the context
This operation renders the operation context in-line so that it can
be sent to the old agent's rabbitmq;  we better include the
rest port in it too, like all other operations do. Otherwise,
the remote agent tries to look up the port in envvars, and that
might be missing (will be missing, for new agents).

Also, use rest_host from our current ctx. It always must be the
current manager, and anyway - we include the rest_token for the
current manager.